requirements.txt
mysql-connector-python
tkcalendar
pandas
matplotlib
